# internal/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: GSE Terminal API
  version: "1.0.0"
  summary: Programmatic access to the Ghana Stock Exchange wire.
  description: |
    Every figure on the GSE Terminal — each tick, close, sentiment score
    and sector aggregate — is available over HTTPS as JSON under the
    `/v1` path.

    ## Authentication

    Most endpoints require a bearer token in the `Authorization` header:

    ```
    Authorization: Bearer ges_live_...
    ```

    Keys are provisioned from **Terminal → Account → API keys** and are
    shown only at creation time. A session cookie works equivalently when
    the call originates from a browser signed into the Terminal.

    ## Rate budget

    - 60 requests per minute, per IP, across the whole surface.
    - LLM endpoints (`/v1/ai-insight`, `/v1/query`) are additionally
      capped at 10 requests per minute, per authenticated user.
    - Exceeding either budget returns `429` with an exponential-backoff
      hint in the response body.

    ## Errors

    Every error response is a JSON body with a single `error` field.
    The HTTP status is authoritative.
  contact:
    name: GSE Terminal
    url: https://your-host/developers
  license:
    name: Proprietary
    url: https://your-host/terms
servers:
  - url: https://{host}/v1
    description: Production
    variables:
      host:
        default: your-host
        description: Your deployment host.
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - bearerApiKey: []
  - sessionCookie: []

# ──────────────────────────────────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────────────────────────────────
paths:
...
  /me/alerts:
    get:
      tags: [Alerts]
      operationId: listAlertRules
      summary: List watchlist alert rules
      description: All rules belonging to the caller — armed and paused. A rule fires at most once before going to paused.
      responses:
        '200':
          description: Array of alert rules.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AlertRule' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Alerts]
      operationId: createAlertRule
      summary: Create a watchlist alert
      description: |
        Requires a verified email on the caller's account. Returns `403`
        if the email is unverified even for Pro callers.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlertRuleInput' }
      responses:
        '201':
          description: The new rule's ID.
          content:
            application/json:
              schema:
                type: object
                required: [id]
                properties:
                  id: { type: integer, format: int64, example: 143 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /me/alerts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Rule ID.
        schema: { type: integer, format: int64, example: 143 }
    patch:
      tags: [Alerts]
      operationId: updateAlertRule
      summary: Update an alert rule
      description: Partial update — any subset of `threshold`, `op`, `metric`, `enabled` may be present.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                threshold: { type: number, example: 1.60 }
                op:        { type: string, enum: ['>', '<', '>=', '<='] }
                metric:    { type: string, enum: [price, rsi, pct_change] }
                enabled:   { type: boolean }
      responses:
        '204': { description: Updated. }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Alerts]
      operationId: deleteAlertRule
      summary: Delete an alert rule
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: ges_live_*
      description: |
        Provision a key from **Terminal → Account → API keys**. The raw
        secret is shown only at creation time.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: JWT-backed session cookie issued by `/login`. Equivalent to a bearer key for browser-originated calls.
  responses:
    BadRequest:
      description: Missing or malformed parameter.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "symbol is required" }
    Unauthorized:
      description: Missing or invalid bearer token / session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "unauthorized" }
    Forbidden:
      description: Authenticated but not Pro/Admin on a Pro-only endpoint, or no verified email for alerts.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "pro access required" }
    NotFound:
      description: No such resource, or not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "not found" }
    Conflict:
      description: Rule limit exceeded, duplicate, or conflicting state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "alert rule limit exceeded" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable error message. }
...
    AlertRuleInput:
      type: object
      required: [symbol, metric, op, threshold]
      properties:
        symbol:    { type: string, example: MTNGH }
        metric:    { type: string, enum: [price, rsi, pct_change], example: price }
        op:        { type: string, enum: ['>', '<', '>=', '<='], example: '>' }
        threshold: { type: number, example: 1.50 }
    AlertRule:
      allOf:
        - $ref: '#/components/schemas/AlertRuleInput'
        - type: object
          required: [id, userId, enabled, fireCount]
          properties:
            id:          { type: integer, format: int64, example: 142 }
            userId:      { type: integer, format: int64, example: 9 }
            enabled:     { type: boolean, example: false }
            lastFiredAt: { type: string, format: date-time, nullable: true, example: '2026-04-14T15:31:02Z' }
            fireCount:   { type: integer, example: 1 }
...